Changing value type for media_count in tags from int to float64 to make it more complete to differents numbers formats in the Instagram API.

+++ instagram/tags.go
@@ -23,8 +23,8 @@ type TagsService struct {
 
 // Tag represents information about a tag object.
 type Tag struct {
-	MediaCount int    `json:"media_count,omitempty"`
-	Name       string `json:"name,omitempty"`
+	MediaCount float64 `json:"media_count,omitempty"`
+	Name       string  `json:"name,omitempty"`
 }
 
 // Get information aout a tag object.
